+++ release-notes.md
@@ -15,6 +15,7 @@ See [log](https://github.com/Pathoschild/SMAPI/compare/1.12...1.13).
 
 For players:
 * SMAPI now has better draw error recovery and detects when the error is irrecoverable.
+* SMAPI now remembers if your game crashed and offers help next time you relaunch.
 
 For mod developers:
 * SMAPI now logs basic context info to simplify troubleshooting, and more detailed logging can be enabled by setting `VerboseLogging: true` in `StardewModdingAPI.config.json`.